Make the project structure
Add a html file and a css file.

Add a recipes folder with three recipes in it.

Add an images folder that contains the hero background image and the recipes result.

Add a navbar and a hero section in the main html file and the styles for them in the css file.